Separated out the tools from the main form, made a factory for creating/holding them, and in general made things more modular. Part 1 of a 2 step process, next is to separate the tool logic from the tool control (kinda like MVC) and to implement a better method of updating when changing settings/ROM data.
