+++ NewSF64Toolkit/NewSF64Toolkit.csproj
@@ -80,6 +80,29 @@
       <DependentUpon>OpenGLControl.cs</DependentUpon>
     </Compile>
     <Compile Include="Program.cs" />
+    <Compile Include="ProgramTools\Controls\HexEditorControl.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="ProgramTools\Controls\HexEditorControl.Designer.cs">
+      <DependentUpon>HexEditorControl.cs</DependentUpon>
+    </Compile>
+    <Compile Include="ProgramTools\Controls\LevelViewerControl.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="ProgramTools\Controls\LevelViewerControl.Designer.cs">
+      <DependentUpon>LevelViewerControl.cs</DependentUpon>
+    </Compile>
+    <Compile Include="ProgramTools\Controls\RomInfoControl.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="ProgramTools\Controls\RomInfoControl.Designer.cs">
+      <DependentUpon>RomInfoControl.cs</DependentUpon>
+    </Compile>
+    <Compile Include="ProgramTools\LevelViewerTool.cs" />
+    <Compile Include="ProgramTools\HexEditorTool.cs" />
+    <Compile Include="ProgramTools\IToolkitMode.cs" />
+    <Compile Include="ProgramTools\RomInfoTool.cs" />
+    <Compile Include="ProgramTools\ToolkitFactory.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="DataStructures\SF64ROM.cs" />
     <Compile Include="StarFoxLevelLoader.cs" />
@@ -94,6 +117,15 @@
     </EmbeddedResource>
     <EmbeddedResource Include="OpenGLControl.resx">
       <DependentUpon>OpenGLControl.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="ProgramTools\Controls\HexEditorControl.resx">
+      <DependentUpon>HexEditorControl.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="ProgramTools\Controls\LevelViewerControl.resx">
+      <DependentUpon>LevelViewerControl.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="ProgramTools\Controls\RomInfoControl.resx">
+      <DependentUpon>RomInfoControl.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
